Build react hook, add peer dependencies, isolate tests
