+++ package.json
@@ -68,5 +68,9 @@
     "axios": "^1.2.3",
     "btoa": "^1.2.1",
     "heresy": "^1.0.4"
+  },
+  "peerDependencies": {
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   }
 }
